+++ Signage/Photo/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 *.png
 dist/
+.env
