+++ Signage/Photo/package.json
@@ -18,6 +18,7 @@
     "@babel/preset-env": "^7.1.6",
     "body-parser": "^1.18.3",
     "cors": "^2.8.5",
+    "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "gyazo-api": "^0.3.1",
     "nodemon": "^1.18.6"
