+++ Signage/Photo/public/index.html
@@ -5,14 +5,17 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <title>Document</title>
+    <link rel="stylesheet" type="text/css" href="style.css" />
   </head>
   <body>
-    <video height="720" width="960" autoplay id="showVideo"></video>
-    <button id="takePicture">take</button> <canvas id="canvas"></canvas>
-    <div class="output">
-      <img id="photo" alt="The screen capture will appear in this box." />
+    <div class="content">
+      <video height="720" width="960" autoplay id="showVideo"></video>
+      <div class="output"><img id="photo" alt="" /></div>
     </div>
+    <div><button id="takePicture">take</button></div>
+    <canvas crossorigin="anonymous" id="canvas"></canvas>
+
     <script src="./jquery.min.js"></script>
-    <script src="../dist/main.js"></script>
+    <script src="../dist/client/main.js"></script>
   </body>
 </html>
