+++ Signage/Photo/src/server/app.js
@@ -4,8 +4,10 @@ import cors from "cors";
 import bodyParser from "body-parser";
 const router = express.Router();
 import fs from "fs";
+import dotenv from "dotenv";
 
-const client = new Gyazo();
+dotenv.load();
+const client = new Gyazo(process.env.ACCESS_TOKEN);
 
 const app = express();
 app.use(bodyParser.json({ limit: "50mb" }));
@@ -22,8 +24,12 @@ app.use(
     console.log("png!");
     const base64Data = req.body.img.replace(/^data:image\/png;base64,/, "");
     const n = Date.now();
-    fs.writeFile("./images/" + n + ".png", base64Data, "base64", err => {
+    fs.writeFile("./images/photo/" + n + ".png", base64Data, "base64", err => {
       console.log(err);
+      client.upload("./images/photo/" + n + ".png", {
+        title: "signage" + n,
+        desc: "upload from connectouch",
+      });
       res.send("書き込み完了");
     });
   })
